refactor(core-api): remove unused chart data types and streamline imports for improved clarity and plugin compatibility
