+++ packages/plugin-awesome/src/charts.ts
@@ -1,17 +1,11 @@
 import {
   ChartDataType,
   ChartType,
-  type TrendChartData,
   type HistoryDataPoint,
   type Statistic,
   type TestResult,
-  type PieChartData,
-  type GeneratedChartData,
-  type GeneratedChartsData,
-  type TrendChartOptions,
-  type PieChartOptions,
 } from "@allurereport/core-api";
-import { type AllureStore, type PluginContext, DEFAULT_CHART_HISTORY_LIMIT, getSeverityTrendData, getStatusTrendData, getPieChartData } from "@allurereport/plugin-api";
+import { type AllureStore, type PluginContext, type TrendChartData, type GeneratedChartData, type GeneratedChartsData, type TrendChartOptions, type PieChartOptions, type PieChartData, DEFAULT_CHART_HISTORY_LIMIT, getSeverityTrendData, getStatusTrendData, getPieChartData } from "@allurereport/plugin-api";
 import { randomUUID } from "crypto";
 import type { AwesomeOptions } from "./model.js";
 import type { AwesomeDataWriter } from "./writer.js";
